+++ pytorch_lightning/trainer/connectors/callback_connector.py
@@ -15,7 +15,8 @@ import os
 from typing import List, Union
 
 from pytorch_lightning.callbacks import Callback, ModelCheckpoint, ProgressBar, ProgressBarBase
-from pytorch_lightning.utilities import rank_zero_warn
+from pytorch_lightning.core.lightning import LightningModule
+from pytorch_lightning.utilities import rank_zero_info, rank_zero_warn
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 
 
@@ -107,6 +108,39 @@ class CallbackConnector:
         for callback in self.trainer.callbacks:
             callback.log = model.log
             callback.log_dict = model.log_dict
+
+    @staticmethod
+    def _attach_model_callbacks(model: LightningModule, trainer) -> None:
+        """
+        Attaches the callbacks defined in the model.
+        If a callback returned by the model's configure_callback method has the same type as one or several
+        callbacks already present in the trainer callbacks list, it will replace them.
+        In addition, all :class:`~pytorch_lightning.callbacks.model_checkpoint.ModelCheckpoint` callbacks
+        will be pushed to the end of the list, ensuring they run last.
+
+        Args:
+            model: A model which may or may not define new callbacks in
+                :meth:`~pytorch_lightning.core.lightning.LightningModule.configure_callbacks`.
+            trainer: The trainer on which the callbacks get attached/merged.
+        """
+        model_callbacks = model.configure_callbacks()
+        if not model_callbacks:
+            return
+        model_callback_types = set(type(c) for c in model_callbacks)
+        trainer_callback_types = set(type(c) for c in trainer.callbacks)
+        override_types = model_callback_types.intersection(trainer_callback_types)
+        if override_types:
+            rank_zero_info(
+                "The following callbacks returned in `LightningModule.configure_callbacks` will override"
+                " existing callbacks passed to Trainer:"
+                f" {', '.join(sorted(t.__name__ for t in override_types))}"
+            )
+        # remove all callbacks with a type that occurs in model callbacks
+        all_callbacks = [c for c in trainer.callbacks if type(c) not in override_types]
+        all_callbacks.extend(model_callbacks)
+        all_callbacks = CallbackConnector._reorder_callbacks(all_callbacks)
+        # TODO: connectors refactor: move callbacks list to connector and do not write Trainer state
+        trainer.callbacks = all_callbacks
 
     @staticmethod
     def _reorder_callbacks(callbacks: List[Callback]) -> List[Callback]: